Parking semi-succesful

+++ stm32-atollic/SelfieF7/Src/Enc/Enc.h
@@ -14,9 +14,9 @@ extern volatile int16_t vleft; //(mm/s)
 extern volatile int16_t vright; //(mm/s)
 extern volatile int16_t vfwd; //(mm/s)
 
-extern volatile int32_t leftRoad; // mm
-extern volatile int32_t rightRoad; // mm
-extern volatile int32_t fwdRoad ; // mm
+extern volatile float leftRoad; // mm
+extern volatile float rightRoad; // mm
+extern volatile float fwdRoad ; // mm
 int16_t wheel_pid(float kp, float ki, float kd, int16_t setfwd);
 
 void StartEncTask(void const * argument);
